Align offload build script with vanilla build script

+++ jlse_offload.sh
@@ -2,8 +2,13 @@
 
 # Make sure all the paths are correct
 
-rm -r build
-rm -r install
+if [ -e build ];then
+  rm -r build
+fi
+
+if [ -e install ];then
+  rm -r install
+fi
 
 MY_PATH=$(pwd)
 
@@ -45,16 +50,10 @@ export BML_TESTING=${BML_TESTING:=yes}
 export CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE:=Release}
 export EXTRA_CFLAGS=${EXTRA_CFLAGS:=""}
 export EXTRA_LINK_FLAGS=${EXTRA_LINK_FLAGS:="-L. -latomic"}
-#export CMAKE_Fortran_FLAGS="-fiopenmp -fopenmp-targets=spir64"
 export CMAKE_Fortran_FLAGS=""
 export CMAKE_C_FLAGS="-fiopenmp -fopenmp-targets=spir64 -D__STRICT_ANSI__"
-#export CMAKE_C_FLAGS="-D__STRICT_ANSI__"
-#export CUDA_TOOLKIT_ROOT_DIR=${CUDATOOLKIT_ROOT}
 
 ./build.sh configure
-cd build
-make -j8
-make test
